+++ variables.tf
@@ -17,6 +17,10 @@ variable "availability_zone" {
   description = "The availability zone to create BIG-IP Next instances in."
   default = "nova"
 }
+variable "destroy" {
+  description = "Whether a destroy operation is being performed."
+  default = false
+}
 variable "mbip_flavor_name" {
   description = "The openstack flavor to use when creating the BIG-IP Next instances."
 }